Remodel: Add Challenge View

+++ app/src/main/java/com/clonect/feeltalk/new_domain/model/question/Question.kt
@@ -3,7 +3,7 @@ package com.clonect.feeltalk.new_domain.model.question
 import java.io.Serializable
 
 data class Question(
-    val index: Int,
+    val index: Long,
     val header: String,
     val body: String,
     val date: String,
